adopt to mac

+++ src/mock/MockData.cpp
@@ -312,18 +312,36 @@ std::shared_ptr<meta::NebulaSchemaProvider> MockData::mockPlayerTagSchema(Object
   schema->addField("age", meta::cpp2::PropertyType::INT64, 0, false);
   schema->addField(
       "playing", meta::cpp2::PropertyType::BOOL, 0, false, ConstantExpression::make(pool, true));
-  schema->addField(
-      "career", meta::cpp2::PropertyType::INT64, 0, false, ConstantExpression::make(pool, 10L));
-  schema->addField(
-      "startYear", meta::cpp2::PropertyType::INT64, 0, false, ConstantExpression::make(pool, 0L));
-  schema->addField(
-      "endYear", meta::cpp2::PropertyType::INT64, 0, false, ConstantExpression::make(pool, 0L));
-  schema->addField(
-      "games", meta::cpp2::PropertyType::INT64, 0, false, ConstantExpression::make(pool, 0L));
-  schema->addField(
-      "avgScore", meta::cpp2::PropertyType::DOUBLE, 0, false, ConstantExpression::make(pool, 0.0));
-  schema->addField(
-      "serveTeams", meta::cpp2::PropertyType::INT64, 0, false, ConstantExpression::make(pool, 0L));
+  schema->addField("career",
+                   meta::cpp2::PropertyType::INT64,
+                   0,
+                   false,
+                   ConstantExpression::make(pool, int64_t(10)));
+  schema->addField("startYear",
+                   meta::cpp2::PropertyType::INT64,
+                   0,
+                   false,
+                   ConstantExpression::make(pool, int64_t(0)));
+  schema->addField("endYear",
+                   meta::cpp2::PropertyType::INT64,
+                   0,
+                   false,
+                   ConstantExpression::make(pool, int64_t(0)));
+  schema->addField("games",
+                   meta::cpp2::PropertyType::INT64,
+                   0,
+                   false,
+                   ConstantExpression::make(pool, int64_t(0)));
+  schema->addField("avgScore",
+                   meta::cpp2::PropertyType::DOUBLE,
+                   0,
+                   false,
+                   ConstantExpression::make(pool, static_cast<double>(0.0)));
+  schema->addField("serveTeams",
+                   meta::cpp2::PropertyType::INT64,
+                   0,
+                   false,
+                   ConstantExpression::make(pool, int64_t(0)));
 
   // Set ttl property
   if (FLAGS_mock_ttl_col) {
@@ -331,7 +349,7 @@ std::shared_ptr<meta::NebulaSchemaProvider> MockData::mockPlayerTagSchema(Object
                      meta::cpp2::PropertyType::INT64,
                      0,
                      false,
-                     ConstantExpression::make(pool, 0L));
+                     ConstantExpression::make(pool, int64_t(0)));
     meta::cpp2::SchemaProp prop;
     prop.set_ttl_col("insertTime");
     prop.set_ttl_duration(FLAGS_mock_ttl_duration);
@@ -375,13 +393,19 @@ std::shared_ptr<meta::NebulaSchemaProvider> MockData::mockServeEdgeSchema(Object
                    meta::cpp2::PropertyType::INT64,
                    0,
                    false,
-                   ConstantExpression::make(pool, 2020L));
-  schema->addField(
-      "endYear", meta::cpp2::PropertyType::INT64, 0, false, ConstantExpression::make(pool, 2020L));
+                   ConstantExpression::make(pool, int64_t(2020)));
+  schema->addField("endYear",
+                   meta::cpp2::PropertyType::INT64,
+                   0,
+                   false,
+                   ConstantExpression::make(pool, int64_t(2020)));
   // only teamCareer filed has no default value and nullable is false
   schema->addField("teamCareer", meta::cpp2::PropertyType::INT64, 0, false);
-  schema->addField(
-      "teamGames", meta::cpp2::PropertyType::INT64, 0, false, ConstantExpression::make(pool, 1L));
+  schema->addField("teamGames",
+                   meta::cpp2::PropertyType::INT64,
+                   0,
+                   false,
+                   ConstantExpression::make(pool, int64_t(1)));
   schema->addField("teamAvgScore",
                    meta::cpp2::PropertyType::DOUBLE,
                    0,
@@ -394,7 +418,7 @@ std::shared_ptr<meta::NebulaSchemaProvider> MockData::mockServeEdgeSchema(Object
                      meta::cpp2::PropertyType::INT64,
                      0,
                      false,
-                     ConstantExpression::make(pool, 0L));
+                     ConstantExpression::make(pool, int64_t(0)));
     meta::cpp2::SchemaProp prop;
     prop.set_ttl_col("insertTime");
     prop.set_ttl_duration(FLAGS_mock_ttl_duration);
@@ -566,7 +590,7 @@ std::shared_ptr<meta::NebulaSchemaProvider> MockData::mockTypicaSchemaV2(ObjectP
                    meta::cpp2::PropertyType::INT64,
                    0,
                    false,
-                   ConstantExpression::make(pool, 20L));
+                   ConstantExpression::make(pool, int64_t(20)));
   schema->addField("col_float", meta::cpp2::PropertyType::FLOAT);
   schema->addField("col_float_null", meta::cpp2::PropertyType::FLOAT, 0, true);
   schema->addField("col_float_default",
